code split & clean up
