+++ src/engine/core.ts
@@ -11,7 +11,7 @@ export abstract class System {
 	 * Update the system.
 	 * @param deltaTime - The time elapsed since the last update call.
 	 */
-	abstract update(deltaTime: number): void | Promise<void>;
+	abstract run(deltaTime: number): void | Promise<void>;
 }
 
 /**
@@ -19,13 +19,13 @@ export abstract class System {
  *
  * This is a variant of an ECS Engine, not sure what to call this, maybe just Systems Pattern
  */
-export class Core {
-	private systems: System[] = [];
+export class Core<S = System> {
+	private systems: S[] = [];
 
 	/**
 	 * Add a system to the Core.
 	 */
-	add<T extends System>(system: T): T {
+	add<T extends S>(system: T): T {
 		this.systems.push(system);
 		return system;
 	}
@@ -34,7 +34,7 @@ export class Core {
 	 * Retrieves all systems of a specified type from the Core.
 	 * @param system - The constructor of the system type.
 	 */
-	get<T extends System>(system: { new (...args: any[]): T }) {
+	get<T extends S>(system: { new (...args: never[]): T }) {
 		return this.systems.filter((s) => s instanceof system) as T[];
 	}
 }
