+++ src/engine/systems/SystemGraph/DependencyGraph.test.ts
@@ -1,13 +1,13 @@
 import { beforeEach, test, describe, expect } from 'bun:test';
-import { DependencyGraph } from './utils';
-import { Core, System } from './core';
+import { Core, System } from '../../core';
+import { DependencyGraph } from './DependencyGraph';
 
 describe('DependencyGraph', () => {
 	class TestSystem extends System {
 		constructor() {
 			super(new Core());
 		}
-		update() {
+		run() {
 			// Do nothing
 		}
 	}
